screen in param

+++ srcs/parser/parser_config.c
@@ -6,7 +6,7 @@
 /*   By: rpinoit <rpinoit@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/26 19:31:10 by rpinoit           #+#    #+#             */
-/*   Updated: 2019/01/29 22:39:08 by rpinoit          ###   ########.fr       */
+/*   Updated: 2019/02/06 13:28:47 by rpinoit          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,6 +53,6 @@ t_config *parser_config(t_json_value *value)
     nerf_double(&config->viewplane.height, 1.0, 3000.0);
     config->ambient = parser_double(getter_by_key(o, "ambient-percentage")) / 100.0;
     nerf_double(&config->ambient, 0.0, 1.0);
-    config->anti_aliasing = parser_double(getter_by_key(o, "anti-aliasing"));
+    config->anti_aliasing = (int)parser_double(getter_by_key(o, "anti-aliasing"));
     return (config);
 }
