+++ srcs/sdl/sdl_pixel.c
@@ -6,7 +6,7 @@
 /*   By: rpinoit <rpinoit@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/04 15:55:47 by rpinoit           #+#    #+#             */
-/*   Updated: 2019/02/04 18:49:32 by rpinoit          ###   ########.fr       */
+/*   Updated: 2019/02/06 13:30:34 by rpinoit          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,6 +17,6 @@
 
 void sdl_pixel(const t_visu *v, const t_color *c, const int x, const int y)
 {
-    SDL_SetRenderDrawColor(v->renderer, c->r, c->g, c->b, 255);
+    SDL_SetRenderDrawColor(v->renderer, (Uint8)c->r, (Uint8)c->g, (Uint8)c->b, 255);
     SDL_RenderDrawPoint(v->renderer, x, y);
 }
